Renamed module and test new functions
